authz: rename PermissionHandler to PermissionPolicy
