+++ packages/backend/package.json
@@ -41,7 +41,7 @@
     "@backstage/plugin-kubernetes-backend": "^0.3.18",
     "@backstage/plugin-kafka-backend": "^0.2.11",
     "@backstage/plugin-permission-backend": "^0.1.0",
-    "@backstage/plugin-permission-handler-simple": "^0.1.0",
+    "@backstage/plugin-permission-policy-simple": "^0.1.0",
     "@backstage/plugin-proxy-backend": "^0.2.13",
     "@backstage/plugin-rollbar-backend": "^0.1.15",
     "@backstage/plugin-scaffolder-backend": "^0.15.13",
