--- plugins/permission-handler-simple/README.md
@@ -1,9 +1,0 @@
-# permission-handler-simple
-
-Welcome to the permission-handler-simple backend plugin!
-
-_This plugin was created through the Backstage CLI_
-
-## Getting started
-
-Your plugin has been added to the example app in this repository, meaning you'll be able to access it by running `yarn start` in the root directory, and then navigating to [/permission-handler-simple](http://localhost:3000/permission-handler-simple).
