+++ plugins/permission-policy-simple/.eslintrc.js
