+++ plugins/permission-policy-simple/api-report.md
@@ -1,4 +1,4 @@
-## API Report File for "@backstage/plugin-permission-handler-simple"
+## API Report File for "@backstage/plugin-permission-policy-simple"
 
 > Do not edit this file. It is a report generated by [API Extractor](https://api-extractor.com/).
 
@@ -6,10 +6,10 @@
 import { BackstageIdentity } from '@backstage/plugin-auth-backend';
 import { EntitiesSearchFilter } from '@backstage/plugin-catalog-backend';
 import { Entity } from '@backstage/catalog-model';
-import { HandlerResult } from '@backstage/plugin-permission-node';
 import { OpaqueAuthorizeRequest } from '@backstage/permission-common';
 import { PermissionCriteria } from '@backstage/permission-common';
-import { PermissionHandler } from '@backstage/plugin-permission-node';
+import { PermissionPolicy } from '@backstage/plugin-permission-node';
+import { PolicyResult } from '@backstage/plugin-permission-node';
 
 // Warning: (ae-missing-release-tag) "isComponentType" is exported by the package, but it is missing a release tag (@alpha, @beta, @public, or @internal)
 //
@@ -21,15 +21,15 @@ export const isComponentType: {
   toQuery(componentTypes: string[]): PermissionCriteria<EntitiesSearchFilter>;
 };
 
-// Warning: (ae-missing-release-tag) "SimplePermissionHandler" is exported by the package, but it is missing a release tag (@alpha, @beta, @public, or @internal)
+// Warning: (ae-missing-release-tag) "SimplePermissionPolicy" is exported by the package, but it is missing a release tag (@alpha, @beta, @public, or @internal)
 //
 // @public (undocumented)
-export class SimplePermissionHandler implements PermissionHandler {
+export class SimplePermissionPolicy implements PermissionPolicy {
   // (undocumented)
   handle(
     request: OpaqueAuthorizeRequest,
     identity?: BackstageIdentity,
-  ): Promise<HandlerResult>;
+  ): Promise<PolicyResult>;
 }
 
 // (No @packageDocumentation comment for this package)
