+++ plugins/permission-policy-simple/src/rules/index.ts
