+++ plugins/permission-policy-simple/src/rules/isComponentType.ts
